Updated the LASSO regularization work
The gradient descent algorithm now works properly but the resulting image is still pretty blurry. I attribute this to using the wrong values for lambda and tau but have instructions for the students to try out other values for them. The one thing I still worry about this is that there aren't many places where I can remove some of the code and have the students fill it in because the majority is used for demonstration purposes and the LASSO part is really just one/two lines.